web_app/requirements.txt
flask==3.0.0
werkzeug==3.0.1
torch==2.2.0
torchvision==0.17.0
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
scikit-learn==1.4.0
joblib==1.3.2
